+++ model_zoo/census_model_sqlflow/wide_and_deep/wide_deep_subclass_keras.py
@@ -8,9 +8,7 @@ from elasticdl_preprocessing.layers.discretization import Discretization
 from elasticdl_preprocessing.layers.hashing import Hashing
 from elasticdl_preprocessing.layers.index_lookup import IndexLookup
 from elasticdl_preprocessing.layers.to_sparse import ToSparse
-from model_zoo.census_model_sqlflow.wide_and_deep.decorators import (
-    model_input_name,
-)
+from elasticdl_preprocessing.utils.decorators import model_input_name
 from model_zoo.census_model_sqlflow.wide_and_deep.feature_configs import (
     INPUT_SCHEMAS,
     age_bucketize,
